cleanup mtarget menu
